web_project/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.103.2",
    "tensorflow==2.15.0; sys_platform != 'darwin' or platform_machine != 'arm64'",
    "tensorflow-macos==2.15.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "tf_keras>=2.15.0",  # Required for TF_USE_LEGACY_KERAS=1
    "pandas==2.0.3",
    "scikit-learn>=1.3.2",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.2",
    "gensim==4.3.3",
    "tabulate==0.9.0",
    "tqdm==4.67.1",
    "networkx==3.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.24",
    "pydantic-settings>=2.12.0",
    "psycopg2-binary>=2.9.11",
    "elasticsearch>=9.2.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "uvicorn>=0.38.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
